+++ netx/dialer/tlsdialer/tlsdialer.go
@@ -7,7 +7,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/ooni/probe-engine/netx/internal/dialer/connx"
+	"github.com/ooni/probe-engine/netx/dialer/connx"
 	"github.com/ooni/probe-engine/netx/internal/errwrapper"
 	"github.com/ooni/probe-engine/netx/modelx"
 )
